chore: clean up schemas

+++ nilcc-agent/src/services/disk.rs
@@ -11,7 +11,7 @@ use tokio::process::Command;
 #[async_trait]
 pub trait DiskService: Send + Sync {
     /// Create a disk at the given path with the given format.
-    async fn create_disk(&self, path: &Path, format: HardDiskFormat, size_gib: u16) -> anyhow::Result<()>;
+    async fn create_disk(&self, path: &Path, format: HardDiskFormat, size_gib: u32) -> anyhow::Result<()>;
 
     /// Create the ISO for an application.
     async fn create_application_iso(&self, path: &Path, spec: IsoSpec) -> anyhow::Result<DockerComposeHash>;
@@ -29,7 +29,7 @@ impl DefaultDiskService {
 
 #[async_trait]
 impl DiskService for DefaultDiskService {
-    async fn create_disk(&self, path: &Path, format: HardDiskFormat, size_gib: u16) -> anyhow::Result<()> {
+    async fn create_disk(&self, path: &Path, format: HardDiskFormat, size_gib: u32) -> anyhow::Result<()> {
         let format = format.to_string();
         let args = ["create", "-f", &format, &path.to_string_lossy(), &format!("{size_gib}G")];
         let output = Command::new(&self.qemu_img_path)
